+++ packages/server/drizzle.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from "drizzle-kit";
+import { ENV } from "./src/lib";
+
+export default defineConfig({
+    schema: './src/db/schema.ts',
+    out: './drizzle',
+    dialect: 'sqlite',
+    dbCredentials: {
+        url: ENV.DATABASE_PATH
+    }
+})