+++ util/deployment/run_remote.sh
@@ -1,13 +1,12 @@
 #!/user/bin/env bash
 set -e
 
-PI_USER="dr_control"
-PI_HOST="192.168.1.5"
-PI_PATH="/home/${PI_USER}/dr_control_project"
+# Get arguments or use defaults
+PI_USER="${1}"
+PI_HOST="${2}"
+PI_PATH="/home/${PI_USER}/water-level-acquisition-system"
 
-
-
-echo "Syncing project to Raspberry Pi..."
+echo "Syncing project wth $PI_USER@$PI_HOST..."
 rsync -avz --delete \
 	--exclude "doc/"  \
 	--exclude ".git/" \
@@ -17,8 +16,10 @@ rsync -avz --delete \
 	--exclude "README.md" \
 	--exclude ".idea/" \
 	--exclude "__pycache__/" \
-	--exclude "util/" \
-	../. ${PI_USER}@${PI_HOST}:${PI_PATH}
+	--exclude "node_modules/" \
+	--exclude "CHANGELOG.md" \
+	--exclude "dist/" \
+	. ${PI_USER}@${PI_HOST}:${PI_PATH}
 
 echo "Running project remotely..."
 ssh -t ${PI_USER}@${PI_HOST} "cd ${PI_PATH}/sampler && /home/${PI_USER}/.local/bin/uv run -m src.sampler"